Move CommentFaker in its own file
