+++ client/my-sites/comments/comment-list/index.jsx
@@ -4,7 +4,7 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { localize } from 'i18n-calypso';
-import { filter, get, keyBy, map, omit, size } from 'lodash';
+import { filter, get, map, size } from 'lodash';
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 
 /**
@@ -18,6 +18,7 @@ import { getNotices } from 'state/notices/selectors';
 import getSiteComments from 'state/selectors/get-site-comments';
 import CommentDetail from 'blocks/comment-detail';
 import CommentDetailPlaceholder from 'blocks/comment-detail/comment-detail-placeholder';
+import CommentFaker from 'blocks/comment-detail/docs/comment-faker';
 import CommentNavigation from '../comment-navigation';
 import EmptyContent from 'components/empty-content';
 
@@ -186,86 +187,6 @@ export class CommentList extends Component {
 		);
 	}
 }
-
-export const CommentFaker = WrappedCommentList => class extends Component {
-	state = {
-		comments: {},
-		isBulkEdit: false,
-	};
-
-	componentWillMount() {
-		if ( this.props.comments.length ) {
-			this.setState( { comments: keyBy( this.props.comments, 'ID' ) } );
-		}
-	}
-
-	componentWillReceiveProps( nextProps ) {
-		if ( ! this.props.comments.length ) {
-			this.setState( { comments: keyBy( nextProps.comments, 'ID' ) } );
-		}
-	}
-
-	deleteCommentPermanently = commentId => this.setState( { comments: omit( this.state.comments, commentId ) } );
-
-	setCommentLike = ( commentId, likeValue ) => {
-		const comment = this.state.comments[ commentId ];
-		// If like changes to true, also approve the comment
-		this.setState( {
-			comments: {
-				...this.state.comments,
-				[ commentId ]: {
-					...comment,
-					i_like: likeValue,
-					status: likeValue ? 'approved' : comment.status,
-				}
-			},
-		} );
-	}
-
-	setCommentStatus = ( commentId, status ) => {
-		const comment = this.state.comments[ commentId ];
-		// If the comment is not approved anymore, also remove the like, otherwise keep its previous value
-		this.setState( {
-			comments: {
-				...this.state.comments,
-				[ commentId ]: {
-					...comment,
-					i_like: 'approved' === status ? comment.i_like : false,
-					status,
-				}
-			},
-		} );
-	}
-
-	toggleCommentLike = commentId => {
-		const comment = this.state.comments[ commentId ];
-		const likeValue = ! comment.i_like;
-		// If like changes to true, also approve the comment
-		this.setState( {
-			comments: {
-				...this.state.comments,
-				[ commentId ]: {
-					...comment,
-					i_like: likeValue,
-					status: likeValue ? 'approved' : comment.status,
-				}
-			},
-		} );
-	}
-
-	render() {
-		return (
-			<WrappedCommentList
-				{ ...this.props }
-				comments={ this.state.comments }
-				deleteCommentPermanently={ this.deleteCommentPermanently }
-				setCommentLike={ this.setCommentLike }
-				setCommentStatus={ this.setCommentStatus }
-				toggleCommentLike={ this.toggleCommentLike }
-			/>
-		);
-	}
-};
 
 const mapStateToProps = ( state, { siteId } ) => {
 	const comments = getSiteComments( state, siteId );
